+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - feat(ssm): preserve ssm runAsEnabled setting and runAsDefaultUser
 - feat(securityHub): added support for AWS Security Hub automation rules
+- feat(config): added support for yaml anchors in json schema
 
 ### Fixed
 - fix(securityhub): fix invalid input exception on batch enabled standards call